Backported spring configuration fix to 4.1. https://issues.apache.org/activemq/browse/AMQ-1276
git-svn-id: https://svn.apache.org/repos/asf/activemq/branches/activemq-4.1@547274 13f79535-47bb-0310-9956-ffa450edef68
