+++ activemq-core/src/test/resources/spring-embedded-xbean.xml
@@ -22,7 +22,7 @@
   xmlns:amq="http://activemq.org/config/1.0"
   xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
   xsi:schemaLocation="http://www.springframework.org/schema/beans http://www.springframework.org/schema/beans/spring-beans-2.0.xsd 
-  http://activemq.org/config/1.0 http://people.apache.org/repository/org.apache.activemq/xsds/activemq-core-4.1-SNAPSHOT.xsd">
+  http://activemq.org/config/1.0 http://activemq.apache.org/snapshot-schema/activemq-core-4.1-SNAPSHOT.xsd">
   
   <!--  lets create an embedded ActiveMQ Broker -->
   <amq:broker useJmx="false" persistent="false">
